requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
python-telegram-bot==20.6